+++ src/duke/core/TaskList.java
@@ -1,3 +1,7 @@
+package duke.core;
+
+import duke.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskList {
