+++ src/duke/task/Task.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class Task{
     private String name;
     private boolean isDone;
