+++ src/duke/task/Todo.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.time.LocalDate;
 
 public class Todo extends Task{
